% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_actionstatus.R
\name{job_actionstatus}
\alias{job_actionstatus}
\title{Utility function to get job action status}
\usage{
job_actionstatus(x)
}
\arguments{
\item{x}{Job ID}
}
\value{
List of action status
}
\description{
Utility function to get job action status
}
