% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_addactions.R
\name{job_addactions}
\alias{job_addactions}
\title{Utility function to add actions to job}
\usage{
job_addactions(x, actions = NULL)
}
\arguments{
\item{x}{Job ID}

\item{actions}{List of actions}
}
\value{
Logical
}
\description{
Appends job with the specified actions.

An action can be specified as an unnamed vector of programs or a vector
named \code{programs}.
}
