% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_addarchive.R
\name{job_addarchive}
\alias{job_addarchive}
\title{Utility function to add archive content to existing job}
\usage{
job_addarchive(x, archive)
}
\arguments{
\item{x}{Job}

\item{archive}{Archive file}
}
\value{
Logical
}
\description{
Updates a job with the specified archive contents. The content is assumed
a zip archive with the root of the zip-file corresponding to the root of
the root of the job work area.
}
