% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_attributes.R
\name{job_attributes}
\alias{job_attributes}
\title{Get job attributes}
\usage{
job_attributes(x)
}
\arguments{
\item{x}{Job ID}
}
\value{
List
}
\description{
Job attributes are additional attributes that is specific to the R compute
service and does not represent thet job attributes maintained by cxlib.
}
