% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_create.R
\name{job_create}
\alias{job_create}
\title{Utility function to create a job}
\usage{
job_create(x, principal = NULL)
}
\arguments{
\item{x}{Job options}
}
\value{
Job ID
}
\description{
Utility function to create a job
}
