% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_exists.R
\name{job_exists}
\alias{job_exists}
\title{Check if a job exists associated with a given principal}
\usage{
job_exists(x, principal = NULL)
}
\arguments{
\item{x}{Job OD}

\item{principal}{Principal name}
}
\value{
Logical
}
\description{
A job is associated with a principal if the job exists and it was created
for the given principal
}
