% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_list.R
\name{job_list}
\alias{job_list}
\title{List jobs for a given principal}
\usage{
job_list(x)
}
\arguments{
\item{x}{Principal}
}
\value{
A list of named elements
}
\description{
List jobs for a given principal
}
