% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_listactions.R
\name{job_listactions}
\alias{job_listactions}
\title{Utility to list job actions}
\usage{
job_listactions(x)
}
\arguments{
\item{x}{Job ID}
}
\value{
List of actions
}
\description{
Utility to list job actions
}
