% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_resultspath.R
\name{job_resultspath}
\alias{job_resultspath}
\title{Utility function to derive path of job results archive}
\usage{
job_resultspath(x)
}
\arguments{
\item{x}{Job ID}
}
\value{
Path to results file
}
\description{
Utility function to derive path of job results archive
}
