% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_status.R
\name{job_status}
\alias{job_status}
\title{Utility function to derive a job's status}
\usage{
job_status(x)
}
\arguments{
\item{x}{Job ID}
}
\value{
Status
}
\description{
The status of a job is represented by one of the key words
\itemize{
\item \code{notstarted} Created and executing first action is yet to start
\item \code{processing} At least one action is currently executing
\item \code{completed} All registered job actions have completed execution
}
}
