% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_info.R
\name{service_info}
\alias{service_info}
\title{Utility function to return standard details for the service R environment}
\usage{
service_info()
}
\value{
Nested list of named entries
}
\description{
Utility function to return standard details for the service R environment
}
