% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_ready.R
\name{service_ready}
\alias{service_ready}
\title{Utility function to return ready status}
\usage{
service_ready()
}
\value{
Logical
}
\description{
Ready status is \code{TRUE} if \code{ps -ef} returns no processes with the identifying
program path string \verb{.../<uuid>/.job/job-*.R}. The process is only visible
within the container that is executing the program.

See work area defined by \link[cxlib]{cxlib_batchjob}.
}
